refactor theme config and support i18n
