+++ src/main/java/run/halo/app/utils/ThemeUtil.java
@@ -1,0 +1,29 @@
+package run.halo.app.utils;
+
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
+import org.springframework.web.servlet.ThemeResolver;
+import org.springframework.web.servlet.support.RequestContextUtils;
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * @author Chuntung Ho
+ */
+public class ThemeUtil {
+
+    /**
+     * Get current active theme from http request attributes.
+     *
+     * @return current active theme
+     */
+    public static String getTheme(HttpServletRequest request) {
+        if (request == null) {
+            request =
+                ((ServletRequestAttributes) (RequestContextHolder.currentRequestAttributes()))
+                    .getRequest();
+        }
+        ThemeResolver themeResolver = RequestContextUtils.getThemeResolver(request);
+        String themeName = themeResolver.resolveThemeName(request);
+        return themeName;
+    }
+}
